+++ packages/cli/test/integration-2.test.ts
@@ -5,7 +5,7 @@ import express from 'express';
 import { createServer } from 'http';
 import { listen } from 'async-listen';
 import { apiFetch } from './helpers/api-fetch';
-import fs, { writeFile, readFile, remove, ensureDir, mkdir } from 'fs-extra';
+import fs, { writeFile, readFile, remove, ensureDir } from 'fs-extra';
 import sleep from '../src/util/sleep';
 import waitForPrompt from './helpers/wait-for-prompt';
 import { execCli } from './helpers/exec';
@@ -1338,7 +1338,7 @@ test('vercel.json configuration overrides in an existing project do not prompt u
   // Step 1. Create a simple static deployment with no configuration.
   // Deployment should succeed and page should display "0"
 
-  await mkdir(path.join(directory, 'public'));
+  await ensureDir(path.join(directory, 'public'));
   await writeFile(path.join(directory, 'public/index.txt'), '0');
 
   // auto-confirm this deployment
@@ -1369,7 +1369,7 @@ test('vercel.json configuration overrides in an existing project do not prompt u
   expect(text).toBe('1\n');
 
   // // Step 3. Do a more complex deployment using a framework this time
-  await mkdir(`${directory}/pages`);
+  await ensureDir(`${directory}/pages`);
   await writeFile(
     `${directory}/pages/index.js`,
     `export default () => 'Next.js Test'`