+++ packages/python/test/fixtures/63-django-whitenoise-with-static-root-staticfiles-dirs/manage.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python
+import os
+import sys
+
+def main():
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "app.settings")
+    from django.core.management import execute_from_command_line
+    execute_from_command_line(sys.argv)
+
+if __name__ == "__main__":
+    main()